fix type

+++ rust/lancedb/src/table.rs
@@ -1286,9 +1286,9 @@ impl NativeTable {
         ivf_params.sample_rate = index.sample_rate as usize;
         ivf_params.max_iters = index.max_iterations as usize;
         let hnsw_params = HnswBuildParams::default()
-            .num_edges(index.m)
-            .max_num_edges(index.m * 2)
-            .ef_construction(index.ef_construction);
+            .num_edges(index.m as usize)
+            .max_num_edges(index.m as usize * 2)
+            .ef_construction(index.ef_construction as usize);
         let sq_params = SQBuildParams {
             sample_rate: index.sample_rate as usize,
             ..Default::default()
